+++ src/ui/test/test_parse_fen.py
@@ -1,43 +1,9 @@
-"""A util to parse a fen string for the UI"""
-
+"""Tests for parsefen"""
 import unittest
+from src.ui.parse_fen import parse_fen
 
-class ChessPiece:
-    """Defines a UI Chess Piece"""
-    def __init__(self, piece_type, color):
-        self.piece_type = piece_type
-        self.color = color
-
-    def __repr__(self):
-        return f"{self.color}{self.piece_type}"
-
-def parse_fen(fen):
-    """Parse a FEN (Forsyth-Edwards Notation) string into an array of pieces"""
-    piece_mapping = {
-        'p': 'pawn', 'r': 'rook', 'n': 'knight', 'b': 'bishop', 'q': 'queen', 'k': 'king',
-        'P': 'pawn', 'R': 'rook', 'N': 'knight', 'B': 'bishop', 'Q': 'queen', 'K': 'king'
-    }
-    color_mapping = {
-        'p': 'black', 'r': 'black', 'n': 'black', 'b': 'black', 'q': 'black', 'k': 'black',
-        'P': 'white', 'R': 'white', 'N': 'white', 'B': 'white', 'Q': 'white', 'K': 'white'
-    }
-
-    rows = fen.split(' ')[0].split('/')
-    board = []
-
-    for row in rows:
-        board_row = []
-        for char in row:
-            if char.isdigit():
-                board_row.extend([None] * int(char))
-            else:
-                piece_type = piece_mapping[char]
-                color = color_mapping[char]
-                board_row.append(ChessPiece(piece_type, color))
-        board.append(board_row)
-
-    return board
-
+class TestParseFen(unittest.TestCase):
+    """Simple unit test for parse_fen"""
 
     def test_initial_position(self):
         """Tests default board state"""
@@ -77,3 +43,5 @@ def parse_fen(fen):
         with self.assertRaises(KeyError):
             parse_fen("9/8/8/8/8/8/8/8")
 
+if __name__ == '__main__':
+    unittest.main()
